src/main/java/com/example/savespecies/exception/ControllerExceptionHandler.java
package com.example.savespecies.exception;

import lombok.extern.log4j.Log4j2;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;

@Log4j2
@ControllerAdvice
public class ControllerExceptionHandler extends ResponseEntityExceptionHandler {
    @ExceptionHandler(DonationNotFoundException.class)
    public ResponseEntity<Void> handleDonationNotFound(Exception exception) {
        log.error("Exception: {}", exception.getMessage(), exception);
        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
    }

    @ExceptionHandler(SpeciesNotFoundException.class)
    public ResponseEntity<Void> handleSpeciesNotFound(Exception exception) {
        log.error("Exception: {}", exception.getMessage(), exception);
        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
    }
}
